v 0.0205c (First optimization issue resolved.)
Found out I was an idiot and didn't clone the inputted color to max and min, so the max and min, along with the new color later on, were fighting over the same object instance.

Now that that is fixed, it looks like there's something wrong with the deviation counting when using two colors for the same hp bar (say hp and shielding).